Moved dialogs into own subdirectory
