+++ app/src/main/java/at/mcbabo/corex/ui/components/SwipableWorkoutExerciseItem.kt
@@ -38,6 +38,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import at.mcbabo.corex.R
 import at.mcbabo.corex.data.entities.WorkoutExercise
+import at.mcbabo.corex.ui.components.dialogs.RecordWeightDialog
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
